requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
amqp==5.0.6
arrow==1.1.1
asgiref==3.4.1
asn1crypto==1.4.0
async-timeout==3.0.1
attrs==21.2.0
backports.zoneinfo==0.2.1
bidict==0.21.2
billiard==3.6.4.0
blessed==1.18.1
cachetools==4.2.2
celery==5.1.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.9
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==35.0.0
decorator==5.0.9
Django==3.2.5
django-picklefield==3.0.1
django-q==1.3.9
djangorestframework==3.12.4
filelock==3.4.0
frozenlist==1.3.0
fsspec==2021.6.1
gcsfs==2021.6.1
google-api-core==1.30.0
google-auth==1.32.1
google-auth-oauthlib==0.4.4
google-cloud-core==1.7.1
google-cloud-storage==1.40.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
google-search-results==2.3.0
googleapis-common-protos==1.53.0
greenlet==1.1.2
gunicorn==20.1.0
huggingface-hub==0.2.1
idna==2.10
joblib==1.1.0
kombu==5.1.0
multidict==5.1.0
netifaces==0.11.0
nltk==3.6.6
numpy==1.21.0
oauthlib==3.1.1
oscrypto==1.2.1
packaging==21.0
pandas==1.3.0
Pillow==8.4.0
prompt-toolkit==3.0.19
protobuf==3.17.
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodomex==3.12.0
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pysnowflake==0.1.3
python-dateutil==2.8.1
python-decouple==3.4
python-engineio==4.2.0
python-socketio==5.3.0
pytz==2021.1
PyYAML==6.0
redis==3.5.3
regex==2021.11.10
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.46
sentencepiece==0.1.96
six==1.16.0
snowflake-connector-python==2.7.3
snowflake-sqlalchemy==1.3.3
SQLAlchemy==1.4.31
sqlparse==0.4.1
threadpoolctl==3.0.0
tokenizers==0.10.3
tornado==6.1
tqdm==4.62.3
transformers==4.14.1
typing-extensions==3.10.0.0
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.1.0
yarl==1.6.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.12.2
